requirements.txt
streamlit
llama-cpp-python
regex
